fix: share scoped Action Scheduler drain service
