+++ inc/Cli/Commands/DrainCommand.php
@@ -9,7 +9,7 @@ namespace DataMachine\Cli\Commands;
 
 use DataMachine\Cli\BaseCommand;
 use DataMachine\Cli\WorkerLock;
-use DataMachine\Core\ActionScheduler\GroupRegistrar;
+use DataMachine\Core\ActionScheduler\ScopedDrainService;
 use WP_CLI;
 
 defined( 'ABSPATH' ) || exit;
@@ -19,11 +19,9 @@ defined( 'ABSPATH' ) || exit;
  */
 class DrainCommand extends BaseCommand {
 
-	private const GROUP = 'data-machine';
+	public const HOOK_BATCH_CHUNK = ScopedDrainService::HOOK_BATCH_CHUNK;
 
-	public const HOOK_BATCH_CHUNK = 'datamachine_pipeline_batch_chunk';
-
-	public const HOOK_EXECUTE_STEP = 'datamachine_execute_step';
+	public const HOOK_EXECUTE_STEP = ScopedDrainService::HOOK_EXECUTE_STEP;
 
 	/**
 	 * Drain due Data Machine actions until empty or a budget is reached.
@@ -116,14 +114,12 @@ class DrainCommand extends BaseCommand {
 	public static function drain( array $options = array() ): array {
 		self::ensureCliMemoryLimit();
 
-		$limit               = max( 0, (int) ( $options['limit'] ?? 0 ) );
-		$batch_size          = max( 1, (int) ( $options['batch_size'] ?? 25 ) );
 		$time_limit          = max( 0, (int) ( $options['time_limit'] ?? 0 ) );
 		$stop_before_timeout = max( 0, (int) ( $options['stop_before_timeout'] ?? 0 ) );
-		$hooks               = self::normalizeHooks( $options['hooks'] ?? null );
+		$service             = new ScopedDrainService();
+		$hooks               = is_array( $options['hooks'] ?? null ) ? $options['hooks'] : null;
 		$job_ids             = self::normalizeJobIds( $options['job_ids'] ?? null );
 		$lane                = self::normalizeLane( $options['lane'] ?? '' );
-		$hooks               = self::hooksForLane( $lane, $hooks );
 		$acquire_lock        = (bool) ( $options['acquire_lock'] ?? true );
 		$lock                = array();
 
@@ -131,7 +127,7 @@ class DrainCommand extends BaseCommand {
 			$lock_ttl = $time_limit > 0 ? $time_limit + max( 60, $stop_before_timeout ) : 600;
 			$lock     = WorkerLock::acquire( (string) ( $options['lock_owner'] ?? self::defaultLockOwner( 'drain', $lane ) ), $lock_ttl, $lane );
 			if ( empty( $lock['acquired'] ) ) {
-				return self::lockedStats( $hooks, $job_ids, $lock, $lane );
+				return self::withLockStatus( $service->lockedStats( $hooks, $job_ids, $lane ), $lock );
 			}
 
 			$lock_token = (string) ( $lock['lock_token'] ?? '' );
@@ -143,77 +139,15 @@ class DrainCommand extends BaseCommand {
 			);
 		}
 
-		$started_at    = time();
-		$before_counts = self::getStatusCounts( $hooks, $job_ids, $lane );
-		$batches       = 0;
-		$warnings      = 0;
-		$stop_reason   = 'empty';
-
 		try {
-			while ( self::getDuePendingCount( $hooks, $job_ids, $lane ) > 0 ) {
-				$stats = self::buildStats( $before_counts, self::getStatusCounts( $hooks, $job_ids, $lane ), $batches, $warnings, $hooks, $job_ids, $stop_reason, $lane );
-				// @phpstan-ignore-next-line
-				if ( self::isMemorySoftLimitReached() ) {
-					$stop_reason = 'memory_limit';
-					break;
-				}
-
-				if ( $limit > 0 && (int) $stats['actions_processed'] >= $limit ) {
-					$stop_reason = 'limit';
-					break;
-				}
-
-				if ( $time_limit > 0 && ( time() - $started_at ) >= $time_limit ) {
-					$stop_reason = 'time_limit';
-					break;
-				}
-
-				if ( $time_limit > 0 && ( $time_limit - ( time() - $started_at ) ) <= $stop_before_timeout ) {
-					$stop_reason = 'timeout_margin';
-					break;
-				}
-
-				$current_batch_size = $batch_size;
-				if ( $limit > 0 ) {
-					$current_batch_size = min( $batch_size, $limit - (int) $stats['actions_processed'] );
-				}
-				if ( $current_batch_size <= 0 ) {
-					$stop_reason = 'limit';
-					break;
-				}
-
-				$due_before    = self::getDuePendingCount( $hooks, $job_ids, $lane );
-				$status_before = self::getStatusCounts( $hooks, $job_ids, $lane );
-				$deadline_at   = 0;
-				if ( $time_limit > 0 ) {
-					$deadline_at = $started_at + max( 0, $time_limit - $stop_before_timeout );
-				}
-				$result = self::runActionSchedulerBatch( $current_batch_size, $hooks, $job_ids, $deadline_at, $lane );
-				++$batches;
-
-				if ( 0 !== (int) $result['return_code'] ) {
-					++$warnings;
-					$message = trim( (string) $result['stderr'] );
-					WP_CLI::warning( '' === $message ? 'Action Scheduler CLI drain failed.' : $message );
-					$stop_reason = 'memory_limit' === (string) $result['stop_reason'] ? 'memory_limit' : 'warning';
-					break;
-				}
-
-				$status_after = self::getStatusCounts( $hooks, $job_ids, $lane );
-				$progress     = (int) ( $result['actions_processed'] ?? self::processedDelta( $status_before, $status_after ) );
-				if ( '' !== (string) $result['stop_reason'] ) {
-					$stop_reason = (string) $result['stop_reason'];
-					break;
-				}
-				if ( 0 === $progress && self::getDuePendingCount( $hooks, $job_ids, $lane ) >= $due_before ) {
-					++$warnings;
-					WP_CLI::warning( 'Drain stopped because Action Scheduler made no observable progress.' );
-					$stop_reason = 'no_progress';
-					break;
-				}
-			}
-
-			$stats = self::buildStats( $before_counts, self::getStatusCounts( $hooks, $job_ids, $lane ), $batches, $warnings, $hooks, $job_ids, $stop_reason, $lane );
+			$stats = $service->drain(
+				$options + array(
+					'execution_context' => 'Data Machine CLI drain',
+					'warning_callback'  => static function ( string $message ): void {
+						WP_CLI::warning( $message );
+					},
+				)
+			);
 			return self::withLockStatus( $stats, $lock );
 		} finally {
 			if ( $acquire_lock ) {
@@ -229,53 +163,11 @@ class DrainCommand extends BaseCommand {
 	 * @return array<string,int|string> Status stats.
 	 */
 	public static function status( array $options = array() ): array {
-		$hooks   = self::normalizeHooks( $options['hooks'] ?? null );
-		$job_ids = self::normalizeJobIds( $options['job_ids'] ?? null );
-		$lane    = self::normalizeLane( $options['lane'] ?? '' );
-		$hooks   = self::hooksForLane( $lane, $hooks );
-		$lock    = WorkerLock::snapshot( null, 600, $lane );
+		$lane   = self::normalizeLane( $options['lane'] ?? '' );
+		$lock   = WorkerLock::snapshot( null, 600, $lane );
+		$status = ( new ScopedDrainService() )->status( $options );
 
-		return array(
-			'due_pending'   => self::getDuePendingCount( $hooks, $job_ids, $lane ),
-			'total_pending' => self::getPendingCount( $hooks, $job_ids, $lane ),
-			'hooks'         => implode( ',', array_keys( self::getStatusCounts( $hooks, $job_ids, $lane ) ) ),
-			'job_ids'       => implode( ',', $job_ids ),
-			'lane'          => $lane,
-		) + self::publicLockStatus( $lock );
-	}
-
-	/**
-	 * Build a lock-skipped drain result.
-	 *
-	 * @param string[]|null                 $hooks   Hook scope.
-	 * @param int[]                         $job_ids Job ID scope.
-	 * @param array<string,int|string|bool> $lock    Lock state.
-	 * @return array<string,int|string> Drain stats.
-	 */
-	private static function lockedStats( ?array $hooks, array $job_ids, array $lock, string $lane = '' ): array {
-		return self::withLockStatus(
-			array(
-				'batches'                    => 0,
-				'batch_chunks'               => 0,
-				'step_executions'            => 0,
-				'completions'                => 0,
-				'failures'                   => 0,
-				'batch_chunk_completions'    => 0,
-				'batch_chunk_failures'       => 0,
-				'step_execution_completions' => 0,
-				'step_execution_failures'    => 0,
-				'actions_processed'          => 0,
-				'other_actions'              => 0,
-				'remaining_pending'          => self::getDuePendingCount( $hooks, $job_ids, $lane ),
-				'total_pending'              => self::getPendingCount( $hooks, $job_ids, $lane ),
-				'warnings'                   => 0,
-				'stop_reason'                => 'locked',
-				'hooks'                      => implode( ',', array_keys( self::getStatusCounts( $hooks, $job_ids, $lane ) ) ),
-				'job_ids'                    => implode( ',', $job_ids ),
-				'lane'                       => $lane,
-			),
-			$lock
-		);
+		return $status + self::publicLockStatus( $lock );
 	}
 
 	/**
@@ -286,18 +178,7 @@ class DrainCommand extends BaseCommand {
 			return;
 		}
 
-		$current = self::memoryLimitBytes();
-		if ( 0 === $current || $current >= $minimum_bytes ) {
-			return;
-		}
-
-		add_filter(
-			'admin_memory_limit',
-			static function () use ( $minimum_bytes ): string {
-				return (string) $minimum_bytes;
-			}
-		);
-		wp_raise_memory_limit( 'admin' );
+		ScopedDrainService::ensureMemoryLimit( $minimum_bytes );
 	}
 
 	/**
@@ -340,523 +221,11 @@ class DrainCommand extends BaseCommand {
 	}
 
 	/**
-	 * Claim and run a batch of due Action Scheduler actions.
-	 *
-	 * @param int           $batch_size Maximum actions to claim.
-	 * @param string[]|null $hooks      Hook scope, or null for all hooks in the group.
-	 * @param int[]         $job_ids    Optional job ID scope.
-	 * @return array{return_code:int,stdout:string,stderr:string,stop_reason:string,actions_processed:int} Result data.
-	 */
-	private static function runActionSchedulerBatch( int $batch_size, ?array $hooks = null, array $job_ids = array(), int $deadline_at = 0, string $lane = '' ): array {
-		$store       = \ActionScheduler_Store::instance();
-		$runner      = \ActionScheduler::runner();
-		$warnings    = array();
-		$claim       = null;
-		$stop_reason = '';
-		$claim_size  = self::claimSizeForScope( $batch_size, $hooks, $job_ids, $lane );
-		$processed   = 0;
-
-		try {
-			GroupRegistrar::ensureDataMachineGroup();
-			self::runActionSchedulerTimeoutCleanup( $store );
-			$claim = $store->stake_claim( $claim_size, null, $hooks ?? array(), self::GROUP );
-		} catch ( \Throwable $throwable ) {
-			return array(
-				'return_code'       => 1,
-				'stdout'            => '',
-				'stderr'            => sprintf( 'Action Scheduler claim failed during drain: %s', $throwable->getMessage() ),
-				'stop_reason'       => 'warning',
-				'actions_processed' => 0,
-			);
-		}
-
-		try {
-			foreach ( $claim->get_actions() as $action_id ) {
-				if ( $deadline_at > 0 && time() >= $deadline_at ) {
-					$stop_reason = 'timeout_margin';
-					break;
-				}
-
-				// @phpstan-ignore-next-line
-				if ( self::isMemorySoftLimitReached() ) {
-					$warnings[] = 'Drain stopped before processing the next action because PHP memory usage reached the soft limit.';
-					break;
-				}
-
-				$action_id = absint( $action_id );
-				if ( $action_id <= 0 || ! self::actionMatchesJobIds( $action_id, $job_ids ) || ! self::actionMatchesLane( $action_id, $lane ) ) {
-					continue;
-				}
-
-				$claimed_action_ids = array_map( 'intval', $store->find_actions_by_claim_id( $claim->get_id() ) );
-				if ( ! in_array( $action_id, $claimed_action_ids, true ) ) {
-					$warnings[] = sprintf( 'Action Scheduler claim %d was lost during drain.', $claim->get_id() );
-					break;
-				}
-
-				try {
-					$runner->process_action( $action_id, 'Data Machine CLI drain' );
-					++$processed;
-				} catch ( \Throwable $throwable ) {
-					$warnings[] = sprintf( 'Action %d failed during drain: %s', $action_id, $throwable->getMessage() );
-				} finally {
-					self::flushRuntimeCache();
-				}
-
-				if ( $processed >= $batch_size ) {
-					break;
-				}
-			}
-		} finally {
-			$store->release_claim( $claim );
-		}
-
-		$stop_reason = self::warningsContainMemoryLimit( $warnings ) ? 'memory_limit' : $stop_reason;
-
-		return array(
-			'return_code'       => empty( $warnings ) ? 0 : 1,
-			'stdout'            => '',
-			'stderr'            => implode( "\n", $warnings ),
-			'stop_reason'       => $stop_reason,
-			'actions_processed' => $processed,
-		);
-	}
-
-	/**
-	 * Determine how many due actions must be claimed to reach a scoped target.
-	 *
-	 * Action Scheduler claims by due order, but does not support filtering a claim
-	 * by serialized job IDs. When a targeted drain is requested, size the claim
-	 * through the first matching job action so the post-claim filter can process it.
-	 *
-	 * @param int           $batch_size Maximum matching actions to process.
-	 * @param string[]|null $hooks      Hook scope, or null for all hooks in the group.
-	 * @param int[]         $job_ids    Optional job ID scope.
-	 */
-	private static function claimSizeForScope( int $batch_size, ?array $hooks, array $job_ids, string $lane ): int {
-		$claim_size = '' === $lane ? $batch_size : max( $batch_size, min( 1000, $batch_size * 20 ) );
-
-		if ( empty( $job_ids ) ) {
-			return $claim_size;
-		}
-
-		return max( $claim_size, self::claimSizeThroughFirstJobAction( $hooks, $job_ids ) );
-	}
-
-	/**
-	 * Count due group actions through the first due action matching a job scope.
-	 *
-	 * @param string[]|null $hooks   Hook scope, or null for all hooks in the group.
-	 * @param int[]         $job_ids Job ID scope.
-	 */
-	private static function claimSizeThroughFirstJobAction( ?array $hooks, array $job_ids ): int {
-		global $wpdb;
-
-		$actions_table = $wpdb->prefix . 'actionscheduler_actions';
-		$groups_table  = $wpdb->prefix . 'actionscheduler_groups';
-		$hook_sql      = self::hookWhereSql( $hooks, $job_ids );
-		$now           = gmdate( 'Y-m-d H:i:s' );
-		$values        = array_merge(
-			array( $actions_table, $groups_table ),
-			$hook_sql['values'],
-			array( self::GROUP, $now )
-		);
-
-		// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber -- Dynamic hook/job scope is constructed from normalized placeholders and values.
-		$target = $wpdb->get_row(
-			$wpdb->prepare(
-				'SELECT a.action_id, a.scheduled_date_gmt, a.priority
-				FROM %i a
-				INNER JOIN %i g ON g.group_id = a.group_id
-				WHERE ' . $hook_sql['sql'] . 'a.status = \'pending\'
-				AND g.slug = %s
-				AND a.scheduled_date_gmt <= %s
-				ORDER BY a.scheduled_date_gmt ASC, a.priority ASC, a.action_id ASC
-				LIMIT 1',
-				$values
-			),
-			ARRAY_A
-		);
-		// phpcs:enable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber
-
-		if ( ! is_array( $target ) ) {
-			return 0;
-		}
-
-		$hook_sql = self::hookWhereSql( $hooks );
-		$values   = array_merge(
-			array( $actions_table, $groups_table ),
-			$hook_sql['values'],
-			array(
-				self::GROUP,
-				(string) $target['scheduled_date_gmt'],
-				(string) $target['scheduled_date_gmt'],
-				(int) $target['priority'],
-				(string) $target['scheduled_date_gmt'],
-				(int) $target['priority'],
-				(int) $target['action_id'],
-			)
-		);
-
-		// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber -- Dynamic hook scope is constructed from normalized placeholders and values.
-		return (int) $wpdb->get_var(
-			$wpdb->prepare(
-				'SELECT COUNT(*)
-				FROM %i a
-				INNER JOIN %i g ON g.group_id = a.group_id
-				WHERE ' . $hook_sql['sql'] . 'a.status = \'pending\'
-				AND g.slug = %s
-				AND (
-					a.scheduled_date_gmt < %s
-					OR (a.scheduled_date_gmt = %s AND a.priority < %d)
-					OR (a.scheduled_date_gmt = %s AND a.priority = %d AND a.action_id <= %d)
-				)',
-				$values
-			)
-		);
-		// phpcs:enable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber
-	}
-
-	/**
-	 * Reset stale claims/running actions before staking a CLI drain claim.
-	 *
-	 * Action Scheduler's own runners call their protected run_cleanup() before
-	 * claiming work. Data Machine claims a group-scoped batch directly, so it must
-	 * perform the same timeout cleanup or old claims can leave due actions counted
-	 * as pending but not claimable.
-	 *
-	 * @param \ActionScheduler_Store $store Action Scheduler store.
-	 */
-	private static function runActionSchedulerTimeoutCleanup( \ActionScheduler_Store $store ): void {
-		if ( ! class_exists( '\ActionScheduler_QueueCleaner' ) ) {
-			return;
-		}
-
-		$time_limit = absint( apply_filters( 'action_scheduler_queue_runner_time_limit', 30 ) );
-		$timeout    = max( 1, $time_limit ) * 10;
-		$cleaner    = new \ActionScheduler_QueueCleaner( $store );
-
-		$cleaner->reset_timeouts( $timeout );
-		$cleaner->mark_failures( $timeout );
-	}
-
-	/**
-	 * Flush in-request object cache state after each CLI-drained action.
-	 */
-	private static function flushRuntimeCache(): void {
-		if ( function_exists( 'wp_cache_supports' ) && wp_cache_supports( 'flush_runtime' ) && function_exists( 'wp_cache_flush_runtime' ) ) {
-			wp_cache_flush_runtime();
-			return;
-		}
-
-		if ( ! function_exists( 'wp_cache_supports' ) && function_exists( 'wp_cache_flush_runtime' ) ) {
-			wp_cache_flush_runtime();
-		}
-	}
-
-	/**
-	 * Whether PHP memory usage is near the hard limit.
-	 */
-	private static function isMemorySoftLimitReached(): bool {
-		$limit = self::memoryLimitBytes();
-		if ( $limit <= 0 ) {
-			return false;
-		}
-
-		$ratio = (float) apply_filters( 'datamachine_cli_drain_memory_soft_limit_ratio', 0.80 );
-		$ratio = max( 0.50, min( 0.95, $ratio ) );
-
-		return memory_get_usage( true ) >= (int) floor( $limit * $ratio );
-	}
-
-	/**
-	 * Return PHP memory_limit in bytes, or 0 for unlimited/unknown.
-	 */
-	private static function memoryLimitBytes(): int {
-		$raw = trim( (string) ini_get( 'memory_limit' ) );
-		if ( '' === $raw || '-1' === $raw ) {
-			return 0;
-		}
-
-		$unit  = strtolower( substr( $raw, -1 ) );
-		$value = (float) $raw;
-		switch ( $unit ) {
-			case 'g':
-				$value *= 1024;
-				// Fall through.
-			case 'm':
-				$value *= 1024;
-				// Fall through.
-			case 'k':
-				$value *= 1024;
-		}
-
-		return max( 0, (int) $value );
-	}
-
-	/**
-	 * Whether a batch warning was caused by the memory soft limit.
-	 *
-	 * @param string[] $warnings Warning messages.
-	 */
-	private static function warningsContainMemoryLimit( array $warnings ): bool {
-		foreach ( $warnings as $warning ) {
-			if ( false !== strpos( $warning, 'memory usage reached the soft limit' ) ) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	/**
-	 * Check whether a claimed action belongs to the requested job scope.
-	 *
-	 * @param int   $action_id Action ID.
-	 * @param int[] $job_ids   Optional job ID scope.
-	 * @return bool True when the action is in scope.
-	 */
-	private static function actionMatchesJobIds( int $action_id, array $job_ids ): bool {
-		if ( empty( $job_ids ) ) {
-			return true;
-		}
-
-		global $wpdb;
-
-		$actions_table = $wpdb->prefix . 'actionscheduler_actions';
-
-		// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching -- This checks the args for an already-claimed Action Scheduler row.
-		$args = (string) $wpdb->get_var(
-			$wpdb->prepare(
-				'SELECT args FROM %i WHERE action_id = %d',
-				$actions_table,
-				$action_id
-			)
-		);
-		// phpcs:enable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching
-
-		foreach ( $job_ids as $job_id ) {
-			if (
-				false !== strpos( $args, '"job_id":' . $job_id . ',' )
-				|| false !== strpos( $args, '"job_id":' . $job_id . '}' )
-				|| false !== strpos( $args, '"parent_job_id":' . $job_id . ',' )
-				|| false !== strpos( $args, '"parent_job_id":' . $job_id . '}' )
-			) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	/**
-	 * Check whether a claimed action belongs to the requested lane.
-	 */
-	private static function actionMatchesLane( int $action_id, string $lane ): bool {
-		if ( '' === $lane ) {
-			return true;
-		}
-
-		global $wpdb;
-
-		$actions_table = $wpdb->prefix . 'actionscheduler_actions';
-
-		// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching -- This checks an already-claimed Action Scheduler row.
-		$row = $wpdb->get_row(
-			$wpdb->prepare(
-				'SELECT hook, args FROM %i WHERE action_id = %d',
-				$actions_table,
-				$action_id
-			),
-			ARRAY_A
-		);
-		// phpcs:enable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching
-
-		if ( ! is_array( $row ) ) {
-			return false;
-		}
-
-		return self::actionRowMatchesLane( (string) ( $row['hook'] ?? '' ), (string) ( $row['args'] ?? '' ), $lane );
-	}
-
-	/**
-	 * Check an Action Scheduler row against a worker lane.
-	 */
-	private static function actionRowMatchesLane( string $hook, string $args, string $lane ): bool {
-		if ( '' === $lane ) {
-			return true;
-		}
-
-		$is_publish = self::HOOK_EXECUTE_STEP === $hook && in_array( self::stepTypeFromExecuteStepArgs( $args ), array( 'ai', 'upsert' ), true );
-		if ( 'publish' === $lane ) {
-			return $is_publish;
-		}
-
-		if ( 'background' === $lane ) {
-			return ! $is_publish;
-		}
-
-		return true;
-	}
-
-	/**
-	 * Resolve an execute-step action's configured step type.
-	 */
-	private static function stepTypeFromExecuteStepArgs( string $args ): string {
-		$decoded = json_decode( $args, true );
-		if ( ! is_array( $decoded ) ) {
-			return '';
-		}
-
-		$job_id       = absint( $decoded['job_id'] ?? 0 );
-		$flow_step_id = (string) ( $decoded['flow_step_id'] ?? '' );
-		if ( $job_id <= 0 || '' === $flow_step_id || ! function_exists( 'datamachine_get_engine_data' ) ) {
-			return '';
-		}
-
-		$engine_data = datamachine_get_engine_data( $job_id );
-		$flow_config = is_array( $engine_data['flow_config'] ?? null ) ? $engine_data['flow_config'] : array();
-		$step_config = is_array( $flow_config[ $flow_step_id ] ?? null ) ? $flow_config[ $flow_step_id ] : array();
-
-		return (string) ( $step_config['step_type'] ?? '' );
-	}
-
-	/**
-	 * Build operator-facing stats.
-	 *
-	 * @param array<string,array<string,int>> $before_counts Counts before drain.
-	 * @param array<string,array<string,int>> $after_counts  Counts after drain.
-	 * @param int                             $batches       Batch count.
-	 * @param int                             $warnings      Warning count.
-	 * @return array<string,int|string> Stats.
-	 */
-	private static function buildStats( array $before_counts, array $after_counts, int $batches, int $warnings, ?array $hooks = null, array $job_ids = array(), string $stop_reason = 'empty', string $lane = '' ): array {
-		$batch_completed   = self::delta( $before_counts, $after_counts, self::HOOK_BATCH_CHUNK, 'complete' );
-		$batch_failed      = self::delta( $before_counts, $after_counts, self::HOOK_BATCH_CHUNK, 'failed' );
-		$step_completed    = self::delta( $before_counts, $after_counts, self::HOOK_EXECUTE_STEP, 'complete' );
-		$step_failed       = self::delta( $before_counts, $after_counts, self::HOOK_EXECUTE_STEP, 'failed' );
-		$total_completed   = self::statusDelta( $before_counts, $after_counts, 'complete' );
-		$total_failed      = self::statusDelta( $before_counts, $after_counts, 'failed' );
-		$total_processed   = $total_completed + $total_failed;
-		$tracked_processed = $batch_completed + $batch_failed + $step_completed + $step_failed;
-
-		return array(
-			'batches'                    => $batches,
-			'batch_chunks'               => $batch_completed + $batch_failed,
-			'step_executions'            => $step_completed + $step_failed,
-			'completions'                => $total_completed,
-			'failures'                   => $total_failed,
-			'batch_chunk_completions'    => $batch_completed,
-			'batch_chunk_failures'       => $batch_failed,
-			'step_execution_completions' => $step_completed,
-			'step_execution_failures'    => $step_failed,
-			'actions_processed'          => $total_processed,
-			'other_actions'              => max( 0, $total_processed - $tracked_processed ),
-			'remaining_pending'          => self::getDuePendingCount( $hooks, $job_ids, $lane ),
-			'total_pending'              => self::getPendingCount( $hooks, $job_ids, $lane ),
-			'warnings'                   => $warnings,
-			'stop_reason'                => $stop_reason,
-			'hooks'                      => implode( ',', self::processedHooks( $before_counts, $after_counts ) ),
-			'job_ids'                    => implode( ',', $job_ids ),
-			'lane'                       => $lane,
-		);
-	}
-
-	/**
-	 * Count status deltas that indicate an action was processed.
-	 *
-	 * @param array<string,array<string,int>> $before_counts Counts before a batch.
-	 * @param array<string,array<string,int>> $after_counts  Counts after a batch.
-	 * @return int Processed action count.
-	 */
-	private static function processedDelta( array $before_counts, array $after_counts ): int {
-		return self::statusDelta( $before_counts, $after_counts, 'complete' ) + self::statusDelta( $before_counts, $after_counts, 'failed' );
-	}
-
-	/**
-	 * Count status deltas across all Data Machine hooks.
-	 *
-	 * @param array<string,array<string,int>> $before_counts Counts before.
-	 * @param array<string,array<string,int>> $after_counts  Counts after.
-	 * @param string                          $status        Action status.
-	 * @return int Non-negative delta.
-	 */
-	private static function statusDelta( array $before_counts, array $after_counts, string $status ): int {
-		$total = 0;
-		foreach ( self::allHooks( $before_counts, $after_counts ) as $hook ) {
-			$total += self::delta( $before_counts, $after_counts, $hook, $status );
-		}
-		return $total;
-	}
-
-	/**
-	 * Get hooks with observed processed deltas.
-	 *
-	 * @param array<string,array<string,int>> $before_counts Counts before.
-	 * @param array<string,array<string,int>> $after_counts  Counts after.
-	 * @return string[] Hook names.
-	 */
-	private static function processedHooks( array $before_counts, array $after_counts ): array {
-		$hooks = array();
-		foreach ( self::allHooks( $before_counts, $after_counts ) as $hook ) {
-			if ( self::delta( $before_counts, $after_counts, $hook, 'complete' ) > 0 || self::delta( $before_counts, $after_counts, $hook, 'failed' ) > 0 ) {
-				$hooks[] = $hook;
-			}
-		}
-		return $hooks;
-	}
-
-	/**
-	 * Get hooks seen in either status snapshot.
-	 *
-	 * @param array<string,array<string,int>> $before_counts Counts before.
-	 * @param array<string,array<string,int>> $after_counts  Counts after.
-	 * @return string[] Hook names.
-	 */
-	private static function allHooks( array $before_counts, array $after_counts ): array {
-		return array_values( array_unique( array_merge( array_keys( $before_counts ), array_keys( $after_counts ) ) ) );
-	}
-
-	/**
-	 * Normalize hook-scope options.
-	 *
-	 * @param mixed $hooks Optional hook list.
-	 * @return string[]|null Hook list, or null for all Data Machine hooks.
-	 */
-	private static function normalizeHooks( mixed $hooks ): ?array {
-		if ( ! is_array( $hooks ) ) {
-			return null;
-		}
-
-		$hooks = array_values(
-			array_unique(
-				array_filter(
-					array_map( 'strval', $hooks ),
-					static fn( string $hook ): bool => '' !== $hook
-				)
-			)
-		);
-
-		return empty( $hooks ) ? null : $hooks;
-	}
-
-	/**
 	 * Normalize a worker lane identifier.
 	 */
 	private static function normalizeLane( mixed $lane ): string {
 		$lane = is_string( $lane ) ? strtolower( trim( $lane ) ) : '';
 		return in_array( $lane, array( 'publish', 'background' ), true ) ? $lane : '';
-	}
-
-	/**
-	 * Restrict hook claims for lanes where hook-level selection is safe.
-	 */
-	private static function hooksForLane( string $lane, ?array $hooks ): ?array {
-		if ( 'publish' === $lane ) {
-			return array( self::HOOK_EXECUTE_STEP );
-		}
-
-		return $hooks;
 	}
 
 	/**
@@ -883,231 +252,5 @@ class DrainCommand extends BaseCommand {
 		}
 
 		return array_values( array_unique( $normalized ) );
-	}
-
-	/**
-	 * Build optional hook WHERE SQL for prepared Action Scheduler queries.
-	 *
-	 * @param string[]|null $hooks Hook scope, or null for all hooks in the group.
-	 * @return array{sql:string,values:string[]} SQL fragment and placeholder values.
-	 */
-	private static function hookWhereSql( ?array $hooks, array $job_ids = array() ): array {
-		$values = array();
-		$sql    = '';
-
-		if ( ! empty( $hooks ) ) {
-			$sql   .= 'a.hook IN (' . implode( ', ', array_fill( 0, count( $hooks ), '%s' ) ) . ') AND ';
-			$values = array_merge( $values, $hooks );
-		}
-
-		if ( ! empty( $job_ids ) ) {
-			$clauses = array();
-			foreach ( $job_ids as $job_id ) {
-				$clauses[] = '(a.args LIKE %s OR a.args LIKE %s OR a.args LIKE %s OR a.args LIKE %s)';
-				$values[]  = '%"job_id":' . $job_id . ',%';
-				$values[]  = '%"job_id":' . $job_id . '}%';
-				$values[]  = '%"parent_job_id":' . $job_id . ',%';
-				$values[]  = '%"parent_job_id":' . $job_id . '}%';
-			}
-			$sql .= '(' . implode( ' OR ', $clauses ) . ') AND ';
-		}
-
-		return array(
-			'sql'    => $sql,
-			'values' => $values,
-		);
-	}
-
-	/**
-	 * Get a status count delta.
-	 *
-	 * @param array<string,array<string,int>> $before_counts Counts before.
-	 * @param array<string,array<string,int>> $after_counts  Counts after.
-	 * @param string                          $hook          Hook name.
-	 * @param string                          $status        Action status.
-	 * @return int Non-negative delta.
-	 */
-	private static function delta( array $before_counts, array $after_counts, string $hook, string $status ): int {
-		return max( 0, ( $after_counts[ $hook ][ $status ] ?? 0 ) - ( $before_counts[ $hook ][ $status ] ?? 0 ) );
-	}
-
-	/**
-	 * Count due pending Data Machine actions.
-	 *
-	 * @return int Due pending count.
-	 */
-	private static function getDuePendingCount( ?array $hooks = null, array $job_ids = array(), string $lane = '' ): int {
-		return self::countActions( true, $hooks, $job_ids, $lane );
-	}
-
-	/**
-	 * Count all pending Data Machine actions.
-	 *
-	 * @return int Pending count.
-	 */
-	private static function getPendingCount( ?array $hooks = null, array $job_ids = array(), string $lane = '' ): int {
-		return self::countActions( false, $hooks, $job_ids, $lane );
-	}
-
-	/**
-	 * Count pending Data Machine actions.
-	 *
-	 * @param bool $due_only Whether to count only due actions.
-	 * @return int Pending count.
-	 */
-	private static function countActions( bool $due_only, ?array $hooks = null, array $job_ids = array(), string $lane = '' ): int {
-		global $wpdb;
-
-		$actions_table = $wpdb->prefix . 'actionscheduler_actions';
-		$groups_table  = $wpdb->prefix . 'actionscheduler_groups';
-		$hook_sql      = self::hookWhereSql( $hooks, $job_ids );
-
-		if ( '' !== $lane ) {
-			return self::countLaneActions( $due_only, $hooks, $job_ids, $lane );
-		}
-
-		if ( $due_only ) {
-			$values = array_merge(
-				array( $actions_table, $groups_table ),
-				$hook_sql['values'],
-				array( self::GROUP, gmdate( 'Y-m-d H:i:s' ) )
-			);
-
-			// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber -- Dynamic hook scope is constructed from normalized placeholders and values.
-			return (int) $wpdb->get_var(
-				$wpdb->prepare(
-					'SELECT COUNT(*)
-					FROM %i a
-					INNER JOIN %i g ON g.group_id = a.group_id
-					WHERE ' . $hook_sql['sql'] . 'a.status = \'pending\'
-					AND g.slug = %s
-					AND a.scheduled_date_gmt <= %s',
-					$values
-				)
-			);
-			// phpcs:enable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber
-		}
-
-		$values = array_merge(
-			array( $actions_table, $groups_table ),
-			$hook_sql['values'],
-			array( self::GROUP )
-		);
-
-		// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber -- Dynamic hook scope is constructed from normalized placeholders and values.
-		return (int) $wpdb->get_var(
-			$wpdb->prepare(
-				'SELECT COUNT(*)
-				FROM %i a
-				INNER JOIN %i g ON g.group_id = a.group_id
-				WHERE ' . $hook_sql['sql'] . 'a.status = \'pending\'
-				AND g.slug = %s',
-				$values
-			)
-		);
-		// phpcs:enable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber
-	}
-
-	/**
-	 * Count pending actions that match a lane.
-	 */
-	private static function countLaneActions( bool $due_only, ?array $hooks, array $job_ids, string $lane ): int {
-		$count = 0;
-		foreach ( self::laneActionRows( $due_only, $hooks, $job_ids ) as $row ) {
-			if ( self::actionRowMatchesLane( (string) ( $row['hook'] ?? '' ), (string) ( $row['args'] ?? '' ), $lane ) ) {
-				++$count;
-			}
-		}
-
-		return $count;
-	}
-
-	/**
-	 * Retrieve Action Scheduler rows for lane filtering.
-	 *
-	 * @return array<int,array<string,string>> Action rows.
-	 */
-	private static function laneActionRows( bool $due_only, ?array $hooks, array $job_ids ): array {
-		global $wpdb;
-
-		$actions_table = $wpdb->prefix . 'actionscheduler_actions';
-		$groups_table  = $wpdb->prefix . 'actionscheduler_groups';
-		$hook_sql      = self::hookWhereSql( $hooks, $job_ids );
-		$status_sql    = "a.status = 'pending'";
-		$date_sql      = $due_only ? 'AND a.scheduled_date_gmt <= %s' : '';
-		$values        = array_merge(
-			array( $actions_table, $groups_table ),
-			$hook_sql['values'],
-			array( self::GROUP )
-		);
-
-		if ( $due_only ) {
-			$values[] = gmdate( 'Y-m-d H:i:s' );
-		}
-
-		// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber -- Dynamic scope is constructed from normalized placeholders and values.
-		$rows = $wpdb->get_results(
-			$wpdb->prepare(
-				'SELECT a.action_id, a.hook, a.status, a.args
-				FROM %i a
-				INNER JOIN %i g ON g.group_id = a.group_id
-				WHERE ' . $hook_sql['sql'] . $status_sql . '
-				AND g.slug = %s
-				' . $date_sql,
-				$values
-			),
-			ARRAY_A
-		);
-		// phpcs:enable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber
-
-		return is_array( $rows ) ? $rows : array();
-	}
-
-	/**
-	 * Get action counts grouped by hook and status.
-	 *
-	 * @return array<string,array<string,int>> Counts by hook and status.
-	 */
-	private static function getStatusCounts( ?array $hooks = null, array $job_ids = array(), string $lane = '' ): array {
-		global $wpdb;
-
-		$actions_table = $wpdb->prefix . 'actionscheduler_actions';
-		$groups_table  = $wpdb->prefix . 'actionscheduler_groups';
-		$hook_sql      = self::hookWhereSql( $hooks, $job_ids );
-		$values        = array_merge(
-			array( $actions_table, $groups_table ),
-			$hook_sql['values'],
-			array( self::GROUP )
-		);
-
-		// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber -- Dynamic hook scope is constructed from normalized placeholders and values.
-		$rows = $wpdb->get_results(
-			$wpdb->prepare(
-				'SELECT a.hook, a.status, COUNT(*) AS count
-				FROM %i a
-				INNER JOIN %i g ON g.group_id = a.group_id
-				WHERE ' . $hook_sql['sql'] . 'g.slug = %s
-				GROUP BY a.hook, a.status',
-				$values
-			)
-		);
-		// phpcs:enable WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.PreparedSQLPlaceholders.ReplacementsWrongNumber
-
-		$counts = array();
-
-		foreach ( $rows as $row ) {
-			$hook   = (string) $row->hook;
-			$status = (string) $row->status;
-			if ( ! isset( $counts[ $hook ] ) ) {
-				$counts[ $hook ] = array(
-					'pending'  => 0,
-					'complete' => 0,
-					'failed'   => 0,
-				);
-			}
-			$counts[ $hook ][ $status ] = (int) $row->count;
-		}
-
-		return $counts;
 	}
 }
